PkgEdition: Interface changed to use strings instead of char*
PMPackage/PMSelection/PMYouPatch: Removed setDataProvider memeber function.
A DataProviderPtr must now be passed to the constructor, it's not desired
to be able to change it, as it's the packages link back to the InstSrcData
it originates from.
Adjusted code in PMYouPatchInfo and InstSrcData_UL to set a NULL DataProviderPtr
to be able to compile until real DataProvider exist.

svn path=/trunk/packagemanager/; revision=4448