- Revised MediaHandler error reporting ang logging. Implemented   MediaCD open/closeTray to use ioctl.
Common logging and error reporting is done by the interface methods
provided by classes MediaHandler/MediaAccess/Mount. Most incidents
will be reported as warning in order not to loose them. It's up to the
functions to decide, whether it ok or a real error.

A certain handler (CD/DISK/NFS/SMB/Curl) should not be in need to log
the requested action and its final result. Class Mount, as used by most
handler logs errors and success, so there's no need to do it twice.

MediaCurl is most probabely the only location where additional logging
is needed to report libcurl errors.

svn path=/trunk/packagemanager/; revision=12336